Move raylib-app integration to bin, revert example
Refactor bin/raylib-libretro.c to use raylib-app's Init/UpdateDrawFrame/
Close callback pattern instead of a manual main() loop:
- AppData carries the LibretroMenu pointer through userData
- Init: opens audio, loads shaders and menu, parses argv for core/game
- UpdateDrawFrame: runs shaders, core, menu update and draw, handles
  F2/F4/F8/F11 hotkeys; returns true (menu drives exit)
- Close: unloads game, core, menu, shaders, audio, frees AppData
- Main: returns App descriptor with FLAG_WINDOW_RESIZABLE | FLAG_VSYNC_HINT

Update bin/CMakeLists.txt to include vendor/raylib-app.
Revert example/ files to their original state (no raylib-app there).
