+++ example/CMakeLists.txt
@@ -4,6 +4,3 @@ add_executable(raylib-libretro-basic
 target_link_libraries(raylib-libretro-basic PUBLIC
     raylib-libretro-static
 )
-target_include_directories(raylib-libretro-basic PRIVATE
-    ../vendor/raylib-app
-)
