+++ example/raylib-libretro-basic.c
@@ -26,83 +26,60 @@
 *
 **********************************************************************************************/
 
-#define RAYLIB_APP_IMPLEMENTATION
-#include "raylib-app.h"
+#include "raylib.h"
 
 #define RAYLIB_LIBRETRO_IMPLEMENTATION
 #include "raylib-libretro.h"
 
-typedef struct {
-    const char* corePath;
-    const char* gamePath;
-} AppData;
-
-bool Init(void** userData, int argc, char** argv) {
+int main(int argc, char* argv[]) {
+    // Ensure proper amount of arguments.
     if (argc <= 1) {
         TraceLog(LOG_ERROR, "Usage: %s <core> [game]", argv[0]);
-        return false;
+        return 1;
     }
 
-    AppData* data = (AppData*)MemAlloc(sizeof(AppData));
-    data->corePath = argv[1];
-    data->gamePath = (argc > 2) ? argv[2] : NULL;
-    *userData = data;
-
+    // Create the window and audio.
+    SetConfigFlags(FLAG_WINDOW_RESIZABLE);
+    InitWindow(800, 600, "raylib-libretro - basic window");
     InitAudioDevice();
 
     // Initialize the given core.
-    if (!InitLibretro(data->corePath)) {
+    if (!InitLibretro(argv[1])) {
         TraceLog(LOG_ERROR, "Failed to initialize libretro core");
         CloseAudioDevice();
-        return false;
+        CloseWindow();
+        return 1;
     }
 
     // Load the given game.
-    if (!LoadLibretroGame(data->gamePath)) {
+    const char* gameFile = (argc > 2) ? argv[2] : NULL;
+    if (!LoadLibretroGame(gameFile)) {
         TraceLog(LOG_ERROR, "Failed to initialize libretro content");
         CloseLibretro();
         CloseAudioDevice();
-        return false;
+        CloseWindow();
+        return 1;
     }
 
-    // Resize the window and set the title to match the loaded core.
-    SetWindowSize((int)GetLibretroWidth(), (int)GetLibretroHeight());
-    SetWindowTitle(GetLibretroName());
+    while (!WindowShouldClose() && !LibretroShouldClose()) {
+        // Run a frame of the core.
+        UpdateLibretro();
 
-    return true;
-}
+        // Render the libretro core.
+        BeginDrawing();
+        {
+            ClearBackground(BLACK);
+            DrawLibretro();
+        }
+        EndDrawing();
+    }
 
-bool UpdateDrawFrame(void* userData) {
-    // Run a frame of the core.
-    UpdateLibretro();
-
-    // Render the libretro core.
-    BeginDrawing();
-        ClearBackground(BLACK);
-        DrawLibretro();
-    EndDrawing();
-
-    return !LibretroShouldClose();
-}
-
-void Close(void* userData) {
     // Unload the game first, and then close the core.
     UnloadLibretroGame();
     CloseLibretro();
 
     CloseAudioDevice();
-    MemFree(userData);
-}
+    CloseWindow();
 
-App Main() {
-    return (App){
-        .title = "raylib-libretro",
-        .width = 800,
-        .height = 600,
-        .init = Init,
-        .update = UpdateDrawFrame,
-        .close = Close,
-        .fps = 60,
-        .configFlags = FLAG_WINDOW_RESIZABLE,
-    };
+    return 0;
 }
